% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FortranwithR-package.R
\docType{package}
\name{FortranwithR-package}
\alias{FortranwithR}
\alias{FortranwithR-package}
\title{FortranwithR: Back to the future: How to use Fortran with R these days?}
\description{
FortranwithR
}
\author{
\strong{Maintainer}: Pawel Cwiek \email{prc.altodato@gmail.com}

Other contributors:
\itemize{
  \item Kurt Schaldemose Hansen (Hansen, Kurt Schaldemose; Vasiljevic, Nikola; Sørensen, Steen Arne (2021). Wind resource data from the tall Risø met mast. Technical University of Denmark. Dataset. https://doi.org/10.11583/DTU.14153204.v1) [contributor]
  \item Nikola Vasiljevic (Hansen, Kurt Schaldemose; Vasiljevic, Nikola; Sørensen, Steen Arne (2021). Wind resource data from the tall Risø met mast. Technical University of Denmark. Dataset. https://doi.org/10.11583/DTU.14153204.v1) [contributor]
  \item Steen Arne Sørensen (Hansen, Kurt Schaldemose; Vasiljevic, Nikola; Sørensen, Steen Arne (2021). Wind resource data from the tall Risø met mast. Technical University of Denmark. Dataset. https://doi.org/10.11583/DTU.14153204.v1) [contributor]
}

}
\keyword{internal}
