% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_sqrt.R
\name{my_sqrt}
\alias{my_sqrt}
\title{My sqrt function}
\usage{
my_sqrt(value)
}
\arguments{
\item{value}{numeric; a number to calculate a root}
}
\value{
resp; a list with results
}
\description{
Step 1: Simple input and output
}
